+++ desktop/cmp/grid/impl/filter/ColumnHeaderFilter.ts
@@ -119,6 +119,6 @@ const switcherButton = hoistCmp.factory<ColumnHeaderFilterModel>(({model, id, ti
         text: title,
         active: activeTabId === id,
         outlined: true,
-        onClick: () => tabContainerModel.activateTab(id)
+        onClick: () => model.activateTab(id)
     });
 });